Limit generated macros to reduce the chance of name collisions.
`$quantities!` is the only remaining generated exported macro. Generated
macros for each quantity no longer exist. This change makes it easier to
have multiple systems containing quantities with the same names.
Quantities in the `units` block of the `system!` macro must always be
prefixed by the quantity's module. This is a breaking change. Prefixing
the module with the `mod` keyword now disables the generation of a
`#[macro_use] pub mod $module;` statement.