Add an `autoconvert` feature
Fixes #54

autoconvert is enabled by default, but when it is disabled, trait
implementations for binary operations do not support mixing different types of
Quantity (of the same measurement units).

This allows performance-sensitive code to avoid the potentially costly
conversions done during these operations even when the same types are on both
sides of the operation.